requirements.txt
subword_nmt==0.3.7
tqdm==4.55.1
websockets==8.1
joblib==1.0.1
sentencepiece==0.1.95
subword_nmt==0.3.7
psutil==5.8.0
pydantic==1.7.3
PyYAML==5.3.1
sacrebleu==1.4.14
websocket_client==0.57.0
